Fixed warning messages

+++ vpr/src/route/bucket.cpp
@@ -110,16 +110,12 @@ Bucket::~Bucket() {
 }
 
 void Bucket::init_heap(const DeviceGrid& grid) {
-    //    vtr::free(heap_);
     delete[] heap_;
     heap_ = nullptr;
 
     heap_size_ = (grid.width() - 1) * (grid.height() - 1);
-    //    heap_ = (BucketItem**)vtr::malloc(heap_size_ * sizeof(BucketItem*));
-    //    memset(heap_, 0, heap_size_ * sizeof(t_heap*));
-
     heap_ = new BucketItem*[heap_size_];
-    for (int i = 0; i < (size_t)heap_size_; i++)
+    for (int i = 0; i < (int)heap_size_; i++)
         heap_[i] = 0;
 
     heap_head_ = std::numeric_limits<size_t>::max();
@@ -151,9 +147,9 @@ void Bucket::expand(size_t required_number_of_buckets) {
     std::vector<BucketItem*> temp(heap_, heap_ + old_size);
     delete[] heap_;
     heap_ = new BucketItem*[heap_size_];
-    for (i = 0; i < old_size; i++)
+    for (i = 0; i < (int)old_size; i++)
         heap_[i] = temp[i];
-    for (i = temp.size(); i < heap_size_; i++)
+    for (i = temp.size(); i < (int)heap_size_; i++)
         heap_[i] = nullptr;
 }
